+++ pingone/types/environment.go
@@ -1,0 +1,6 @@
+package types
+
+func EnvironmentTypeList() []string {
+
+	return []string{"PRODUCTION", "SANDBOX"}
+}