+++ pingone/types/schemaAttribute.go
@@ -1,0 +1,6 @@
+package types
+
+func SchemaAttributeTypeList() []string {
+
+	return []string{"STRING", "JSON", "BOOLEAN", "COMPLEX"}
+}