+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Release (2022-07-16-2)
+
+* **Feature** `github.com/patrickcping/pingone-go-sdk-v2` : v0.2.0
+    * New `pingone` package with aggregated client initialisation
+    * New `environment`, `schemaAttribute` and `role` helper types
+
 # Release (2022-07-16)
 
 Initial release - rebasing versions to reflect module stability